Strealine attacking and remove still human
I was pretty attached to the idea of only human and still human being
distinct, but it really didn't make a lot of sense. If people (myself
included), want to generalise the game, of course they can. But at the
moment, what I want to use it for, and what it really has been built
with in mind (with respect to the flimsy characters and whatnot) is
near-future semi-hard sci-fi, so really that is what it should be built
for.

I also made impact more straight forward and the attack process more
symetric, after actually trying the game for only the third or so time
and realising the existing one was hard and weird to explain (especially
with the possibly negative impact).

Oh, and everything should be modifiers now. And also much more 10s
based.