Complete first draft of new character creation
In addition to some other bits that came along with it as I thought of
them. It is much nicer to not feel constrained by the layout, I have a
lot more space to be clear, and to give examples. But, damn, it is going
to be very hard to feel like I can reduce this to anything like the size
it was before and still be doing it justice.